SERVERLESS-2326 Implement context creation in the non-Lambda runner (#17)
We've "merged" the Lambda and the non-Lambda runtimes with the goal of being able to benefit from Lambda's context capabilities. However, the runner does a lot of more things that don't really make sense outside of a "Lambda compatibility layer" context. Those are:

1. Forcing a Promise return for functions that have more than 1 argument.
2. Supporting a 3rd callback parameter.
3. Mangling the input event if it looks like an HTTP event constructed by the controller.
4. Supporting Lambda-specific capabilities like `callbackWaitsForEmptyEventLoop`.

This backs some of the merge out again and implements context-mapping "natively" in the non-Lambda runner to make it behave like before but pass an additional context parameter.