Downgrade npm to version 8 to fix UID remapping issues (#19)
Recently, the Node.js docker image was updated from 18.13.0 to 18.14.0 which also caused an npm upgrade from 8.19.3 to 9.3.1.

The 9.x series of npm has made changes in that it stops mucking with the ownership of the installed node_modules. That's causing us trouble right now. A discussion of the matter can be found in https://github.com/npm/cli/issues/5889.

Downgrading npm to 8.x unwedges the issue for now.