SERVERLESS-2833 Write sentinels straight to stdout/stderr
This changes the sentinels to be written directly to FD1 (stdout) and FD2 (stderr) to avoid users tampering with how the sentinels get written and thus break our expectations towards how the sentinels are written.

This has been an issue with customers when they used packages like log-timestamp to patch the console functions to output timestamps as well.